🔧 Chore: Separate bootstrap configuration from app entry point
Refactor app initialization by:
- Creating a new `bootstrap.ts` file to handle module aliases and environment configuration
- Simplifying the `app.ts` file by removing bootstrap-related imports and setup
- Improving separation of concerns in application startup process
